refactor: test if ~/.ssh folder exist
If applied, this commit will...

Add a if exist test for the ~/.ssh folder. If folder doesn't exist creates a new one

Why is this change being made?

When the user has not made any ssh configuration (probably a fresh install of OS) the install process stops on .ssh configuration step